Added validation in SportsServices, added IntegrationTests and refactored code.

+++ src/main/kotlin/pt/isel/ls/sports/services/SportsServices.kt
@@ -1,4 +1,4 @@
-package pt.isel.ls.sports
+package pt.isel.ls.sports.services
 
 import pt.isel.ls.sports.api.routers.users.CreateUserResponse
 import pt.isel.ls.sports.data.SortOrder
@@ -12,6 +12,19 @@ import pt.isel.ls.sports.errors.SportsError
 class SportsServices(private val db: SportsDatabase) {
 
     /**
+     * Gets the user's unique identifier associate with the [token]
+     *
+     * @param token user token
+     *
+     * @return user's unique identifier associated with the [token]
+     *
+     * @throws SportsError.invalidCredentials if a user with the [token] was not found
+     */
+    private fun authenticate(token: String) = runCatching {
+        db.getUID(token)
+    }.getOrElse { throw SportsError.invalidCredentials() }
+
+    /**
      * Creates a new user in the database.
      *
      * @param name user's name
@@ -20,6 +33,15 @@ class SportsServices(private val db: SportsDatabase) {
      * @return a createUser response (user's token and the user's unique identifier)
      */
     fun createNewUser(name: String, email: String): CreateUserResponse {
+        if (!User.isValidName(name))
+            throw SportsError.invalidArgument("Name must be between ${User.MIN_NAME_LENGTH} and ${User.MAX_NAME_LENGTH} characters")
+
+        if (!User.isValidEmail(email))
+            throw SportsError.invalidArgument("Invalid email")
+
+        if (db.hasUserWithEmail(email))
+            throw SportsError.invalidArgument("Email already in use")
+
         val uid = db.createNewUser(name, email)
         val token = db.createUserToken(uid)
 
@@ -34,6 +56,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return user object
      */
     fun getUser(uid: Int): User {
+        if (!isValidId(uid))
+            throw SportsError.invalidArgument("User id must be positive")
+
         return db.getUser(uid)
     }
 
@@ -61,6 +86,8 @@ class SportsServices(private val db: SportsDatabase) {
      */
     fun createNewRoute(token: String, startLocation: String, endLocation: String, distance: Double): Int {
         val uid = authenticate(token)
+        if (!Route.isValidDistance(distance))
+            throw SportsError.invalidArgument("Distance must be positive")
 
         return db.createNewRoute(
             startLocation,
@@ -78,6 +105,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return the route object
      */
     fun getRoute(rid: Int): Route {
+        if (!isValidId(rid))
+            throw SportsError.invalidArgument("Route id must be positive")
+
         return db.getRoute(rid)
     }
 
@@ -102,7 +132,13 @@ class SportsServices(private val db: SportsDatabase) {
     fun createNewSport(token: String, name: String, description: String?): Int {
         val uid = authenticate(token)
 
-        return db.createNewSport(name, description ?: "", uid)
+        if (!Sport.isValidName(name))
+            throw SportsError.invalidArgument("Name must be between ${Sport.MIN_NAME_LENGTH} and ${Sport.MAX_NAME_LENGTH} characters")
+
+        if (description != null && !Sport.isValidDescription(description))
+            throw SportsError.invalidArgument("Description must be between ${Sport.MIN_DESCRIPTION_LENGTH} and ${Sport.MAX_DESCRIPTION_LENGTH} characters")
+
+        return db.createNewSport(uid, name, description)
     }
 
     /**
@@ -113,6 +149,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return the sport object
      */
     fun getSport(sid: Int): Sport {
+        if (!isValidId(sid))
+            throw SportsError.invalidArgument("Sport id must be positive")
+
         return db.getSport(sid)
     }
 
@@ -139,7 +178,19 @@ class SportsServices(private val db: SportsDatabase) {
     fun createNewActivity(token: String, date: String, duration: String, sid: Int, rid: Int?): Int {
         val uid = authenticate(token)
 
-        return db.createNewActivity(date, duration, uid, sid, rid)
+        if (!Activity.isValidDate(date))
+            throw SportsError.invalidArgument("Date must be in the format yyyy-mm-dd")
+
+        if (!Activity.isValidDuration(duration))
+            throw SportsError.invalidArgument("Duration must be in the format hh:mm:ss.fff")
+
+        if (!isValidId(sid))
+            throw SportsError.invalidArgument("Sport id must be positive")
+
+        if (rid != null && !isValidId(rid))
+            throw SportsError.invalidArgument("Route id must be positive")
+
+        return db.createNewActivity(uid, date, duration, sid, rid)
     }
 
     /**
@@ -150,6 +201,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return the activity object
      */
     fun getActivity(aid: Int): Activity {
+        if (!isValidId(aid))
+            throw SportsError.invalidArgument("Activity id must be positive")
+
         return db.getActivity(aid)
     }
 
@@ -158,7 +212,16 @@ class SportsServices(private val db: SportsDatabase) {
      *
      * @param aid activity's unique identifier
      */
-    fun deleteActivity(aid: Int) {
+    fun deleteActivity(token: String, aid: Int) {
+        val uid = authenticate(token)
+        val activity = db.getActivity(aid)
+
+        if (uid != activity.uid)
+            throw SportsError.forbidden("You are not allowed to delete this activity")
+
+        if (!isValidId(aid))
+            throw SportsError.invalidArgument("Activity id must be positive")
+
         return db.deleteActivity(aid)
     }
 
@@ -170,6 +233,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return list of activities of a sport
      */
     fun getSportActivities(sid: Int): List<Activity> {
+        if (!isValidId(sid))
+            throw SportsError.invalidArgument("Sport id must be positive")
+
         return db.getSportActivities(sid)
     }
 
@@ -181,6 +247,9 @@ class SportsServices(private val db: SportsDatabase) {
      * @return list of activities made from a user
      */
     fun getUserActivities(uid: Int): List<Activity> {
+        if (!isValidId(uid))
+            throw SportsError.invalidArgument("User id must be positive")
+
         return db.getUserActivities(uid)
     }
 
@@ -196,27 +265,20 @@ class SportsServices(private val db: SportsDatabase) {
      *
      * @return list of activities identifiers
      */
-    fun getActivities(sid: Int, orderBy: String, date: String?, rid: Int?, limit: Int?, skip: Int?): List<Activity> {
-        val order =
-            if (orderBy == "ascending")
-                SortOrder.ASCENDING
-            else
-                SortOrder.DESCENDING
+    fun getActivities(
+        sid: Int,
+        orderBy: String,
+        date: String?,
+        rid: Int?,
+        limit: Int?,
+        skip: Int?
+    ): List<Activity> {
+        if (!isValidId(sid))
+            throw SportsError.invalidArgument("Sport id must be positive")
 
-        return db.getActivities(sid, order, date, rid)
-            .run { return@run subList(fromIndex = skip ?: 0, toIndex = (limit ?: lastIndex) + 1) }
+        val order = SortOrder.parse(orderBy)
+            ?: throw SportsError.invalidArgument("Order by must be either ascending or descending")
+
+        return db.getActivities(sid, order, date, rid, skip, limit)
     }
-
-    /**
-     * Gets the user's unique identifier associate with the [token]
-     *
-     * @param token user token
-     *
-     * @return user's unique identifier associated with the [token]
-     *
-     * @throws SportsError.invalidCredentials if a user with the [token] was not found
-     */
-    private fun authenticate(token: String) = runCatching {
-        db.getUID(token)
-    }.getOrElse { throw SportsError.invalidCredentials() }
 }
